backend/WEB_serv/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic-settings==2.2.1
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
redis==5.0.4
httpx==0.27.0
python-dotenv==1.0.1
loguru==0.7.2
